refactor: rename smart contract to KyodoRegistry
